+++ metarunlog/schedulers.py
@@ -14,8 +14,9 @@ import getpass
 class AbstractScheduler:
     def __init__(self, expDir, jobList, resourceName, resourceProp):
         self.expDir  = expDir
-        self.jobList = jobList
-        self.nextJobIx = 0
+        self.jobList = []
+        self.runningJobs = None # override
+        for job in jobList: self.addNewJob(job)
         self.resourceName = resourceName
         self.resourceProp = resourceProp
         self.sleepInterval = resourceProp.get('sleepInterval', 5)
@@ -33,28 +34,45 @@ class AbstractScheduler:
 
     def main(self):
         while not all(job.isFinished() for job in self.jobList):
-            finishedJobs  = self.popFinishedJobs()
-            resourceAvail = self.nextAvailableResource()
-            if resourceAvail is not None and not all(job.isStarted() for job in self.jobList):
-                nextJob = self.jobList[self.nextJobIx]
-                print('Start job {} on resource {}'.format(str(nextJob), self.fmtResource(resourceAvail)))
-                self.startJob(nextJob, resourceAvail)
-                self.nextJobIx += 1
+            finishedJobs = self.popFinishedJobs()
+            nextResource = self.nextAvailableResource()
+            nextJob      = self.nextJob()
+            if nextResource is not None and nextJob:
+                print('Start job {} on resource {}'.format(str(nextJob), self.fmtResource(nextResource)))
+                self.startJob(nextJob, nextResource)
             else:
                 self.printStatus()
                 self.sleep()
         self.printStatus()
         print('\nScheduler finished execution.')
 
+    def addNewJob(self, job):
+        self.jobList.append(job)
+
+    def nextJob(self):
+        """ Get next job: first the jobs that have to be resumed, then jobs that haven't started yet. """
+        for job in self.jobList:
+            if job.started and not job.finished and not job in self.runningJobs:
+                return job #to be resumed
+        for job in self.jobList:
+            if not job.started:
+                return job
+        return None
+
     def popFinishedJobs(self):
-        raise Exception("override")
+        finished         = [job for job in self.runningJobs if (job and not job.isRunning())]
+        self.runningJobs = [job if (job and job.isRunning()) else None for job in self.runningJobs]
+        return finished
 
     def nextAvailableResource(self):
-        """ 
-        returns the next available resource in a format understood by startJob,
-        or the None object if no resource available.
         """
-        raise Exception("override")
+        returns index of first available slot if a slot is available,
+        or the None object if all resource slots are running a job.
+        """
+        try:
+            return self.runningJobs.index(None)
+        except ValueError:
+            return None
 
     def startJob(self, resource):
         raise Exception("override")
@@ -74,21 +92,27 @@ class AbstractScheduler:
                 elif job.finished:
                     msg += 'FINISHED'
                 else:
-                    msg += 'RUNNING'
+                    if job in self.runningJobs:
+                        msg += 'RUNNING'
+                    else:
+                        msg += 'TO BE RESUMED'
             else:
                 msg += 'NOT STARTED'
             print(msg)
         print('---')
 
     def terminate(self):
-        print("{} Asking running jobs to terminate.".format(self.__class__.__name__))
-        for job in self.jobList:
-            job.terminate()
+        rjobs = [job for job in self.runningJobs if job is not None]
+        if rjobs:
+            print("{} Asking running jobs to terminate.".format(self.__class__.__name__))
+            for job in rjobs:
+                job.terminate()
 
     def __del__(self):
         # kill all jobs before removing lockfile 
-        ##TODO check that this actually kills children, see http://stackoverflow.com/questions/4789837/how-to-terminate-a-python-subprocess-launched-with-shell-true
-        print("{} Sending SIGKILL to remaining jobs".format(self.__class__.__name__))
+        rjobs = [job for job in self.runningJobs if job is not None]
+        if rjobs:
+            print("{} Sending SIGKILL to remaining jobs".format(self.__class__.__name__))
         del self.jobList[:]
         print("{} remove .mrl.running".format(self.__class__.__name__))
         os.remove(join(self.expDir, '.mrl.running'))
@@ -99,31 +123,16 @@ class localScheduler(AbstractScheduler):
         TODO: support multiple devices OR do not set the lockfile in runJobs
         """
         AbstractScheduler.__init__(self, expDir, jobList, resourceName, resourceProp)
-        self.runningJob = None
+        self.runningJobs = [None] # one slot
 
-    def popFinishedJobs(self):
-        if self.runningJob and not self.runningJob.isRunning():
-            finished = self.runningJob
-            self.runningJob = None
-            return [finished]
-        else:
-            return []
-
-    def nextAvailableResource(self):
-        """ 
-        returns 'local' if it is available, 
-        or the None object if a job is running.
-        """
-        return 'local' if not self.runningJob else None
-        #TODO have multiple devices in config file
-
-    def startJob(self, job, resource):
-        assert not self.runningJob, "self.runningJob not None"
+    def startJob(self, job, slotIx):
+        assert self.runningJobs[slotIx] is None, "Host {} has job {} running".\
+                format(self.fmtResource(slotIx), str(self.runningJobs[slotIx]))
         job.startLocal()
-        self.runningJob = job
+        self.runningJobs[slotIx] = job
 
     def fmtResource(self, resource):
-        return resource
+        return 'local'
 
 class sshScheduler(AbstractScheduler):
     def __init__(self, expDir, jobList, resourceName, resourceProp):
@@ -131,30 +140,15 @@ class sshScheduler(AbstractScheduler):
         self.hosts = [(host, int(dev)) for host,dev in self.resourceProp['hosts']]
         self.runningJobs = [None for slot in self.hosts]
 
-    def popFinishedJobs(self):
-        finished         = [job for job in self.runningJobs if (job and not job.isRunning())]
-        self.runningJobs = [job if (job and job.isRunning()) else None for job in self.runningJobs]
-        return finished
-
-    def nextAvailableResource(self):
-        """
-        returns index into self.hosts of first available slot if a host is available,
-        or the None object if all hosts are running a job.
-        """
-        try:
-            return self.runningJobs.index(None)
-        except ValueError:
-            return None
-
-    def startJob(self, job, resourceIx):
-        assert self.runningJobs[resourceIx] is None, "Host {} has job {} running".\
-                format(self.fmtResource(resourceIx), str(self.runningJobs[resourceIx]))
-        host, device = self.hosts[resourceIx]
+    def startJob(self, job, slotIx):
+        assert self.runningJobs[slotIx] is None, "Host {} has job {} running".\
+                format(self.fmtResource(slotIx), str(self.runningJobs[slotIx]))
+        host, device = self.hosts[slotIx]
         job.startSsh(host, device, self.sshPass, self.copyFiles)
-        self.runningJobs[resourceIx] = job
+        self.runningJobs[slotIx] = job
 
-    def fmtResource(self, resourceIx):
-        return '{} [device {}]'.format(*self.hosts[resourceIx])
+    def fmtResource(self, slotIx):
+        return '{} [device {}]'.format(*self.hosts[slotIx])
 
 class pbsScheduler(AbstractScheduler):
     """ Note, pbsScheduler and job.startPbs() is not implemented for monitoring yet.
@@ -164,20 +158,12 @@ class pbsScheduler(AbstractScheduler):
     def __init__(self, expDir, jobList, resourceName, resourceProp):
         AbstractScheduler.__init__(self, expDir, jobList, resourceName, resourceProp)
         self.host = self.resourceProp['host']
-        self.runningJob = None
-        # TODO maxConcurrent with pbs monitoring
+        self.runningJobs = [None] # only one slot since it finishes right away
 
-    def popFinishedJobs(self):
-        finished = self.runningJob # jobs finish right away
-        self.runningJob = None
-        return [finished]
-
-    def nextAvailableResource(self):
-        return self.resourceName # just submit all of 'em
-
-    def startJob(self, job, resource):
+    def startJob(self, job, slotIx):
+        assert self.runningJobs[slotIx] is None, "Pbs scheduler should not have a job running (should finish right away)"
         job.startPbs(self.host, self.sshPass, self.copyFiles, self.resourceProp.get('qsubHeader', []))
-        self.runningJob = job
+        self.runningJobs[slotIx] = job
 
     def fmtResource(self, resource):
         return resource
